chage package name to org.drools...
