+++ drools-game-engine-services/src/test/java/org/drools/game/services/endpoint/test/cmds/ClearPlayerInventoryCommand.java
@@ -14,32 +14,21 @@
  * limitations under the License.
  */
 
-package com.drools.game.services.endpoint.test.cmds;
+package org.drools.game.services.endpoint.test.cmds;
 
 import org.drools.game.core.api.BaseCommand;
 import org.drools.game.core.api.Context;
 import org.drools.game.model.api.Player;
 
-public class NotifyAllViaChatCommand extends BaseCommand<Void> {
+public class ClearPlayerInventoryCommand extends BaseCommand<Void> {
 
-    private String message;
-
-    public NotifyAllViaChatCommand( Player player, String message ) {
+    public ClearPlayerInventoryCommand( Player player ) {
         super( player );
-        this.message = message;
     }
 
     @Override
     public Void execute( Context ctx ) {
         throw new UnsupportedOperationException( "Not supported yet." ); //To change body of generated methods, choose Tools | Templates.
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage( String message ) {
-        this.message = message;
     }
 
 }
